+++ src/LanguageProvider.test.ts
@@ -1,4 +1,10 @@
-import { AbstractLabelOperator, AbstractQuery, DataSourceInstanceSettings, DataSourcePluginMeta, PluginType } from '@grafana/data';
+import {
+  AbstractLabelOperator,
+  AbstractQuery,
+  DataSourceInstanceSettings,
+  DataSourcePluginMeta,
+  PluginType,
+} from '@grafana/data';
 
 import LanguageProvider from './LanguageProvider';
 import { DataSource } from './DataSource';
@@ -32,8 +38,8 @@ describe('transform abstract query to a LogScale query', () => {
     const result = instance.importFromAbstractQuery(abstractQuery);
 
     expect(result).toEqual({
-      lsql: "label1=\"value1\"\n| label2 != \"value2\"\n| label3 = *value3*\n| label4 != *value4*",
-      repository: "repo",
+      lsql: 'label1="value1"\n| label2 != "value2"\n| label3 = *value3*\n| label4 != *value4*',
+      repository: 'repo',
       refId: abstractQuery.refId,
     } as LogScaleQuery);
   });
@@ -42,15 +48,13 @@ describe('transform abstract query to a LogScale query', () => {
     const instance = new LanguageProvider(datasource);
     const abstractQuery: AbstractQuery = {
       refId: 'bar',
-      labelMatchers: [
-        { name: '__name__', operator: AbstractLabelOperator.Equal, value: 'repo' },
-      ],
+      labelMatchers: [{ name: '__name__', operator: AbstractLabelOperator.Equal, value: 'repo' }],
     };
     const result = instance.importFromAbstractQuery(abstractQuery);
 
     expect(result).toEqual({
-      lsql: "",
-      repository: "repo",
+      lsql: '',
+      repository: 'repo',
       refId: abstractQuery.refId,
     } as LogScaleQuery);
   });
@@ -62,8 +66,8 @@ describe('transform abstract query to a LogScale query', () => {
 
     expect(result).toEqual({
       refId: abstractQuery.refId,
-      lsql: "",
-      repository: "",
+      lsql: '',
+      repository: '',
     } as LogScaleQuery);
   });
 });
@@ -82,6 +86,7 @@ export function createDataSource(
       type: PluginType.datasource,
     } as DataSourcePluginMeta,
     name: 'test-data-source',
+    readOnly: false,
     type: 'type',
     uid: 'uid',
     access: 'proxy',