+++ src/components/LogScaleQueryEditor.test.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { render, waitFor, screen, act } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { getMockDatasource } from './__fixtures__/datasource';
+import { mockDatasource } from './__fixtures__/datasource';
 import { LogScaleQueryEditor, Props } from './LogScaleQueryEditor';
 
 const getDefaultProps = (): Props => {
   const props: Props = {
-    datasource: getMockDatasource(),
+    datasource: mockDatasource(),
     onChange: jest.fn(),
     onRunQuery: jest.fn(),
     query: {