+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 1.2.0
+
+- Bug: Issue where users were unable to select default repository is fixed.
+
+## 1.1.0
+
+- Minimum Grafana required version is now **9.5.0**
+- Logs in explore view can be filtered by a value or filtered out by a value.
+- The settings UI has been overhauled to use the new Grafana form and authentication components.
+
 ## 1.0.1
 
 - Bug: TLS option are now correctly passed to the LogScale client.